update the prefix search test.
