+++ dora/core/client/fs/src/test/java/alluxio/client/file/cache/LocalCacheManagerTest.java
@@ -43,6 +43,7 @@ import alluxio.file.ByteArrayTargetBuffer;
 import alluxio.file.NettyBufTargetBuffer;
 import alluxio.file.ReadTargetBuffer;
 import alluxio.network.protocol.databuffer.DataFileChannel;
+import alluxio.uri.UfsUrl;
 import alluxio.util.CommonUtils;
 import alluxio.util.WaitForOptions;
 import alluxio.util.io.BufferUtils;
@@ -55,8 +56,10 @@ import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.DefaultFileRegion;
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Assume;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -1201,6 +1204,62 @@ public final class LocalCacheManagerTest {
     byte[] bytes = new byte[PAGE1.length];
     buf.readBytes(bytes);
     assertArrayEquals(PAGE1, bytes);
+  }
+
+  @Ignore("Remove this annotation once OS worker supports the prefix search.")
+  @Test
+  public void searchPagesByPrefix() throws Exception {
+    mCacheManager = createLocalCacheManager();
+
+    String scheme = "file:///";
+    String prefix1 = scheme + "prefix1";
+    String prefix2 = scheme + "prefix2";
+    String prefix3 = scheme + "q";
+    String prefix4 = scheme + "~";
+
+    int fileNum = 10;
+    cachePrefixPages(prefix1, fileNum);
+    cachePrefixPages(prefix2, fileNum);
+    cachePrefixPages(prefix3, fileNum);
+    cachePrefixPages(prefix4, fileNum);
+
+    UfsUrl rootPrefix = UfsUrl.createInstance(scheme);
+
+    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix1));
+    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix2));
+    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix3));
+    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix4));
+  }
+
+  private void cachePrefixPages(String prefix, int fileNum) {
+    for (int i = 0; i < fileNum; i++) {
+      String fileId = prefix + "_" + i;
+      PageId pageId = new PageId(fileId, 0);
+      mCacheManager.put(pageId, PAGE1);
+    }
+  }
+
+  private void checkPrefixSearch(UfsUrl rootUrl, UfsUrl prefixUrl) {
+    List<PageInfo> rootResult = mCacheManager.getPageInfoByPrefix(rootUrl);
+    Set<PageInfo> rootResultSet = new HashSet<>(rootResult);
+    Assert.assertEquals(rootResult.size(), rootResultSet.size());
+    Assert.assertFalse(rootResultSet.isEmpty());
+
+    List<PageInfo> prefixSearchResult = mCacheManager.getPageInfoByPrefix(prefixUrl);
+    Set<PageInfo> prefixSearchSet = new HashSet<>(prefixSearchResult);
+    Assert.assertEquals(prefixSearchResult.size(), prefixSearchSet.size());
+    Assert.assertFalse(prefixSearchSet.isEmpty());
+
+    for (PageInfo p : prefixSearchResult) {
+      Assert.assertTrue(p.getUfsUrl().toString().startsWith(prefixUrl.toString()));
+    }
+    for (PageInfo p : rootResultSet) {
+      if (prefixSearchSet.contains(p)) {
+        Assert.assertTrue(p.getUfsUrl().toString().startsWith(prefixUrl.toString()));
+      } else {
+        Assert.assertFalse(p.getUfsUrl().toString().startsWith(prefixUrl.toString()));
+      }
+    }
   }
 
   /**
