+++ dora/core/server/worker/src/test/java/alluxio/worker/dora/PagedDoraWorkerTest.java
@@ -24,7 +24,6 @@ import alluxio.client.file.FileSystemContext;
 import alluxio.client.file.cache.CacheManager;
 import alluxio.client.file.cache.CacheManagerOptions;
 import alluxio.client.file.cache.PageId;
-import alluxio.client.file.cache.PageInfo;
 import alluxio.client.file.cache.PageMetaStore;
 import alluxio.conf.Configuration;
 import alluxio.conf.PropertyKey;
@@ -54,9 +53,7 @@ import alluxio.metrics.MetricKey;
 import alluxio.metrics.MetricsSystem;
 import alluxio.security.authorization.Mode;
 import alluxio.underfs.UfsStatus;
-import alluxio.uri.UfsUrl;
 import alluxio.util.io.BufferUtils;
-import alluxio.util.io.PathUtils;
 import alluxio.wire.WorkerIdentity;
 import alluxio.worker.block.BlockMasterClientPool;
 
@@ -65,7 +62,6 @@ import com.google.common.util.concurrent.ListenableFuture;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -78,9 +74,7 @@ import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -1012,60 +1006,6 @@ public class PagedDoraWorkerTest {
       int size = end - start;
       buffer.writeBytes(mBuffer, start, size);
       return size;
-    }
-  }
-
-  @Ignore("Remove this annotation once OS worker supports the prefix search.")
-  @Test
-  public void searchPagesByPrefix() throws Exception {
-    String prefix1 = "prefix1";
-    String prefix2 = "prefix2";
-    String prefix3 = "q";
-    String prefix4 = "~";
-
-    UfsUrl rootPrefix = PathUtils.convertUfsPathToUfsUrl(
-        mTestFolder.getRoot().getAbsolutePath() + "/");
-
-    int fileNum = 10;
-    createFilesWithPrefix(prefix1, fileNum);
-    createFilesWithPrefix(prefix2, fileNum);
-    createFilesWithPrefix(prefix3, fileNum);
-    createFilesWithPrefix(prefix4, fileNum);
-
-    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix1));
-    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix2));
-    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix3));
-    checkPrefixSearch(rootPrefix, rootPrefix.join(prefix4));
-  }
-
-  private void checkPrefixSearch(UfsUrl rootUrl, UfsUrl prefixUrl) {
-    List<PageInfo> rootResult = mWorker.getPageInfoByPrefix(rootUrl);
-    Set<PageInfo> rootResultSet = new HashSet<>(rootResult);
-    Assert.assertEquals(rootResult.size(), rootResultSet.size());
-
-    List<PageInfo> prefixSearchResult = mWorker.getPageInfoByPrefix(prefixUrl);
-    Set<PageInfo> prefixSearchSet = new HashSet<>(prefixSearchResult);
-    Assert.assertEquals(prefixSearchResult.size(), prefixSearchSet.size());
-
-    for (PageInfo p : prefixSearchResult) {
-      Assert.assertTrue(p.getUfsUrl().toString().startsWith(prefixUrl.toString()));
-    }
-    for (PageInfo p : rootResultSet) {
-      if (prefixSearchSet.contains(p)) {
-        Assert.assertTrue(p.getUfsUrl().toString().startsWith(prefixUrl.toString()));
-      } else {
-        Assert.assertFalse(p.getUfsUrl().toString().startsWith(prefixUrl.toString()));
-      }
-    }
-  }
-
-  private void createFilesWithPrefix(String prefix, int fileNum)
-      throws AccessControlException, IOException, ExecutionException, InterruptedException,
-      TimeoutException {
-    for (int i = 0; i < fileNum; i++) {
-      File f = mTestFolder.newFile(prefix + "_" + i);
-      Files.write(f.toPath(), mFileContent.getBytes());
-      loadFileData(f.getPath());
     }
   }
 }
